query_sessions/20260409_140913/pms/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
pytest-asyncio==0.21.1
pytest-mock==3.12.0
